fix containers-filter typing and rename file
